+++ apps/expo-go/ios/Exponent-Bridging-Header.h
@@ -2,6 +2,13 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
+#import <ExpoModulesCore/ExpoModulesCore.h>
+#import <ExpoModulesCore-Swift.h>
+#import <EXUpdatesInterface-Swift.h>
+#import <EXUpdates-Swift.h>
+#import <ExpoScreenOrientation-Swift.h>
+#import <EXManifests-Swift.h>
+
 #import "ExpoKit.h"
 #import "EXKernel.h"
 #import "EXRootViewController.h"
@@ -9,6 +16,9 @@
 #import "EXVersionManagerObjC.h"
 #import "EXScopedModuleRegistryDelegate.h"
 #import "EXScopedModuleRegistryAdapter.h"
+#import "ExpoAppInstance.h"
 #import "EXAppState.h"
 #import "EXDisabledDevLoadingView.h"
 #import "EXStatusBarManager.h"
+#import "EXKernelDevKeyCommands.h"
+#import "EXClientReleaseType.h"