+++ apps/expo-go/ios/Exponent/Kernel/Views/EXAppViewController.mm
@@ -23,13 +23,13 @@
 #import "EXUtil.h"
 
 #import <EXSplashScreen/EXSplashScreenService.h>
-#import <React/RCTUtils.h>
 #import <ExpoModulesCore/EXModuleRegistryProvider.h>
 
+#import <React/RCTUtils.h>
 #import <React/RCTAppearance.h>
+#import <React/RCTDevSettings.h>
 
 #import <RNScreens/RNSScreenWindowTraits.h>
-
 
 #define EX_INTERFACE_ORIENTATION_USE_MANIFEST 0
 
@@ -55,7 +55,7 @@ NS_ASSUME_NONNULL_BEGIN
   <EXReactAppManagerUIDelegate, EXAppLoaderDelegate, EXErrorViewDelegate, EXAppLoadingCancelViewDelegate>
 
 @property (nonatomic, assign) BOOL isLoading;
-@property (atomic, assign) BOOL isBridgeAlreadyLoading;
+@property (atomic, assign) BOOL isHostAlreadyLoading;
 @property (nonatomic, weak) EXKernelAppRecord *appRecord;
 @property (nonatomic, strong) EXErrorView *errorView;
 @property (nonatomic, strong) NSTimer *tmrAutoReloadDebounce;
@@ -234,7 +234,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)refresh
 {
   self.isLoading = YES;
-  self.isBridgeAlreadyLoading = NO;
+  self.isHostAlreadyLoading = NO;
   [self _invalidateRecoveryTimer];
   [_appRecord.appLoader request];
 }
@@ -242,7 +242,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)reloadFromCache
 {
   self.isLoading = YES;
-  self.isBridgeAlreadyLoading = NO;
+  self.isHostAlreadyLoading = NO;
   [self _invalidateRecoveryTimer];
   [_appRecord.appLoader requestFromCache];
 }
@@ -276,15 +276,15 @@ NS_ASSUME_NONNULL_BEGIN
 {
 }
 
-- (void)_rebuildBridge
+- (void)_rebuildHost
 {
-  if (!self.isBridgeAlreadyLoading) {
-    self.isBridgeAlreadyLoading = YES;
+  if (!self.isHostAlreadyLoading) {
+    self.isHostAlreadyLoading = YES;
     dispatch_async(dispatch_get_main_queue(), ^{
       [self _overrideUserInterfaceStyleOf:self];
       [self _overrideAppearanceModuleBehaviour];
       [self _invalidateRecoveryTimer];
-      [self.appRecord.appManager rebuildBridge];
+      [self.appRecord.appManager rebuildHost];
     });
   }
 }
@@ -418,7 +418,7 @@ NS_ASSUME_NONNULL_BEGIN
     UIView *rootView = self.view;
     UIView *splashScreenView = [provider createSplashScreenView];
     self.managedSplashScreenController = [[EXManagedAppSplashScreenViewController alloc] initWithRootView:rootView
-                                                                                                 splashScreenView:splashScreenView];
+                                                                                splashScreenView:splashScreenView];
     [splashScreenService showSplashScreenFor:self
                                      options:EXSplashScreenDefault
                       splashScreenController:self.managedSplashScreenController
@@ -453,7 +453,7 @@ NS_ASSUME_NONNULL_BEGIN
   if ([EXKernel sharedInstance].browserController) {
     [[EXKernel sharedInstance].browserController addHistoryItemWithUrl:appLoader.manifestUrl manifest:manifest];
   }
-  [self _rebuildBridge];
+  [self _rebuildHost];
 }
 
 - (void)appLoader:(EXAbstractLoader *)appLoader didLoadBundleWithProgress:(EXLoadingProgress *)progress
@@ -471,7 +471,7 @@ NS_ASSUME_NONNULL_BEGIN
     BOOL forceRTL = [self _readForcesRTLFromManifest:_appRecord.appLoader.manifest];
     [EXTextDirectionController setRTLPreferences:supportsRTL :forceRTL];
   }
-  [self _rebuildBridge];
+  [self _rebuildHost];
   if (self->_appRecord.appManager.status == kEXReactAppManagerStatusBridgeLoading) {
     [self->_appRecord.appManager appLoaderFinished];
   }
@@ -598,11 +598,11 @@ NS_ASSUME_NONNULL_BEGIN
 {
   [super viewWillTransitionToSize:size withTransitionCoordinator:coordinator];
 
-  __weak typeof(self) weakSelf = self;
+  __weak EXAppViewController *weakSelf = self;
 
   // Update after the transition ends, this ensures that the trait collection passed to didUpdateDimensionsEvent is already updated
   [coordinator animateAlongsideTransition:^(id<UIViewControllerTransitionCoordinatorContext> _Nonnull context) {
-    __strong __typeof(self) strongSelf = weakSelf;
+    __strong EXAppViewController *strongSelf = weakSelf;
 
     if (!strongSelf) {
       return;