+++ packages/@expo/cli/static/template/[...rsc]+api.ts
@@ -1,10 +1,10 @@
 import { getRscMiddleware } from '@expo/server/build/middleware/rsc';
 import { renderRscAsync } from 'expo-router/build/rsc/middleware';
 
-import { join } from 'node:path';
+import { resolve } from 'node:path';
 
 // Target the `dist/server` directory.
-const distFolder = join(__dirname, '../../..');
+const distFolder = resolve('./');
 
 const rscMiddleware = getRscMiddleware({
   config: {},