+++ packages/@expo/metro-config/src/serializer/jsOutput.ts
@@ -30,6 +30,7 @@ export type JsOutput = {
     hasCjsExports?: boolean;
 
     readonly reconcile?: ReconcileTransformSettings;
+    readonly reactServerReference?: string;
     readonly reactClientReference?: string;
     readonly expoDomComponentReference?: string;
   };
@@ -69,7 +70,7 @@ export type ReconcileTransformSettings = {
   };
   collectDependenciesOptions: CollectDependenciesOptions;
 
-  unstable_dependencyMapReservedName?: string;
+  unstable_dependencyMapReservedName?: string | null;
   optimizationSizeLimit?: number;
   unstable_disableNormalizePseudoGlobals?: boolean;