+++ packages/expo-doctor/src/utils/doctorConfig.ts
@@ -1,4 +1,7 @@
+import { ExpoConfig, PackageJSONConfig } from '@expo/config';
+
 import { env } from './env';
+import { gteSdkVersion } from './versions';
 
 /**
  * Get the doctor config from the package.json.
@@ -39,7 +42,8 @@ export function getReactNativeDirectoryCheckExcludes(pkg: any) {
   });
 }
 
-export function getReactNativeDirectoryCheckEnabled(pkg: any) {
+export function getReactNativeDirectoryCheckEnabled(exp: ExpoConfig, pkg: PackageJSONConfig) {
+  const isEnabledByDefault = gteSdkVersion(exp, '52.0.0');
   const pkgJsonConfigSetting = getDoctorConfig(pkg).reactNativeDirectoryCheck?.enabled;
 
   if (env.EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK !== null) {
@@ -52,7 +56,7 @@ export function getReactNativeDirectoryCheckEnabled(pkg: any) {
     return env.EXPO_DOCTOR_ENABLE_DIRECTORY_CHECK;
   }
 
-  return pkgJsonConfigSetting ?? false;
+  return pkgJsonConfigSetting ?? isEnabledByDefault;
 }
 
 export function getReactNativeDirectoryCheckListUnknownPackagesEnabled(pkg: any) {