+++ packages/expo-doctor/CHANGELOG.md
@@ -6,15 +6,26 @@
 
 ### 🎉 New features
 
-- Show more info on network errors, allow overriding of network errors ([#31926](https://github.com/expo/expo/pull/31926) by [@keith-kurak](https://github.com/keith-kurak))
+### 🐛 Bug fixes
 
+### 💡 Others
+
+## 1.11.1 — 2024-10-15
+
+### 🐛 Bug fixes
+
+- Remove `--minify` from ncc for now, in order to prevent log spam on errors.
+
+## 1.11.0 — 2024-10-15
+
+### 🎉 New features
+
+- Show more info on network errors, allow overriding of network errors ([#31926](https://github.com/expo/expo/pull/31926) by [@keith-kurak](https://github.com/keith-kurak))
 - Add support for `.easignore` files when performing project validations. ([#31334](https://github.com/expo/expo/pull/31334) by [@betomoedano](https://github.com/betomoedano))
 
 ### 🐛 Bug fixes
 
 - fix conditional on iOS non-CNG check ([#31920](https://github.com/expo/expo/pull/31920) by [@keith-kurak](https://github.com/keith-kurak))
-
-### 💡 Others
 
 ## 1.10.1 — 2024-08-28